Added prepareForExport function in utils
