+++ src/utils/prepareForExport.js
@@ -1,0 +1,31 @@
+function prepareForExport (component) {
+  // Declare install function executed by Vue.use()
+  component.install = function (Vue, settings) {
+    if (component.installed) {
+      return;
+    }
+    component.installed = true;
+
+    const name = settings.altName ? settings.altName : component.name;
+    Vue.component(name, component);
+  };
+
+  // Create module definition for Vue.use()
+  const plugin = {
+    install: component.install,
+  };
+
+  // Auto-install when vue is found (eg. in browser via <script> tag)
+  let GlobalVue = null;
+  if (typeof window !== 'undefined') {
+    GlobalVue = window.Vue;
+  } else if (typeof global !== 'undefined') {
+    GlobalVue = global.Vue;
+  }
+  if (GlobalVue) {
+    GlobalVue.use(plugin);
+  }
+
+}
+
+export default prepareForExport;
